Apply suggestions in PR #81 - Move new scripts from .../maintel to .../maintel/tma - Fix typo in RandomWalk docstring - In random_walk, convert generator into async function without arguments - In serpent_walk, convert generator into async function without arguments - In serpent_walk, convert generate_azel_sequence into static method - Use setUpClass instead of setUp in tests
