+++ python/lsst/ts/externalscripts/data/scripts/maintel/tma/random_walk.py
@@ -1,6 +1,7 @@
-# This file is part of ts_externalcripts.
+#!/usr/bin/env python
+# This file is part of ts_externalscripts
 #
-# Developed for the Rubin Observatory Telescope and Site System.
+# Developed for the LSST Telescope and Site Systems.
 # This product includes software developed by the LSST Project
 # (https://www.lsst.org).
 # See the COPYRIGHT file at the top-level directory of this distribution
@@ -17,8 +18,9 @@
 # GNU General Public License for more details.
 #
 # You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from .make_comcam_calibrations import *
-from .track_target_sched import *
-from .warmup_hexapod import *
+import asyncio
+
+from lsst.ts.externalscripts.maintel.tma import RandomWalk
+
+asyncio.run(RandomWalk.amain())
