+++ python/lsst/ts/externalscripts/maintel/tma/__init__.py
@@ -1,7 +1,6 @@
-#!/usr/bin/env python
-# This file is part of ts_externalscripts
+# This file is part of ts_externalcripts.
 #
-# Developed for the LSST Telescope and Site Systems.
+# Developed for the Rubin Observatory Telescope and Site System.
 # This product includes software developed by the LSST Project
 # (https://www.lsst.org).
 # See the COPYRIGHT file at the top-level directory of this distribution
@@ -18,9 +17,7 @@
 # GNU General Public License for more details.
 #
 # You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import asyncio
-
-from lsst.ts.externalscripts.maintel import SerpentWalk
-
-asyncio.run(SerpentWalk.amain())
+from .random_walk import *
+from .serpent_walk import *
